+++ .zombienet-tests/0002-checkSync.ts
@@ -1,0 +1,19 @@
+import { connect } from "./utils"
+
+export async function run(nodeName: string, networkInfo: any) {
+  const { chainHead } = await connect(nodeName, networkInfo, "100")
+  let count = 0
+
+  await new Promise(async (resolve, reject) => {
+    const chainHeadFollower = chainHead(
+      true,
+      (event) => {
+        if (event.event === "finalized" && ++count === 2) {
+          resolve(chainHeadFollower.unfollow())
+        }
+      },
+      reject,
+    )
+  })
+  return count
+}